fix(eslint-plugin-query): Reorganise internal modules
