+++ packages/eslint-plugin-query/src/rules.ts
@@ -1,0 +1,5 @@
+import * as exhaustiveDeps from './rules/exhaustive-deps.rule'
+
+export const rules = {
+  [exhaustiveDeps.name]: exhaustiveDeps.rule,
+}
