+++ packages/eslint-plugin-query/src/rules/exhaustive-deps.rule.ts
@@ -1,7 +1,7 @@
 import { AST_NODE_TYPES } from '@typescript-eslint/utils'
-import { ASTUtils } from '../../utils/ast-utils'
-import { createRule } from '../../utils/create-rule'
-import { uniqueBy } from '../../utils/unique-by'
+import { ASTUtils } from '../utils/ast-utils'
+import { createRule } from '../utils/create-rule'
+import { uniqueBy } from '../utils/unique-by'
 import { ExhaustiveDepsUtils } from './exhaustive-deps.utils'
 import type { TSESLint } from '@typescript-eslint/utils'
