--- packages/eslint-plugin-query/src/rules/index.ts
@@ -1,5 +1,0 @@
-import * as exhaustiveDeps from './exhaustive-deps/exhaustive-deps.rule'
-
-export const rules = {
-  [exhaustiveDeps.name]: exhaustiveDeps.rule,
-}
